+++ Projektarbeit/src/QuicksortFirstPivot.java
@@ -5,7 +5,7 @@
  * @author Hermann Witte
  */
 public class QuicksortFirstPivot implements Sortieralgorithmus{
-    private int vergleiche = 0;
+    private long vergleiche = 0;
     private double zeit;
     private int schreibzugriffe = 0;
     private int length;
@@ -80,7 +80,7 @@ public class QuicksortFirstPivot implements Sortieralgorithmus{
      * @return Gibt die Anzahl der vergleiche, welcher der Algorithmus benötigt zurück.
      */
     @Override
-    public int getVergleiche() {
+    public long getVergleiche() {
         return vergleiche;
     }
